Separate Solution class and NamedSolution class into two separate files.
